requirements.txt
PyYAML==3.11
requests==2.10.0